change About folder
