+++ client/src/App.js
@@ -6,6 +6,7 @@ import ResetLocation from './helpers/ResetLocation';
 import LoginModal from './components/login/LoginModal';
 import Home from './routes/landing/Home';
 import Contact from './routes/contact/Contact';
+import About from './routes/about/About';
 
 function App() {
   const [productQuantity, setProductQuantity] = useState(0);
@@ -78,6 +79,7 @@ function App() {
       <Routes>
         <Route path='/' element={<Home />} />
         <Route path='/contact' element={<Contact />} />
+        <Route path='/about' element={<About />} />
       </Routes>
       <Footer />
     </BrowserRouter>
